Add testing utilities, add sym world wrapper for exposing protected functions for unit testing, continue updating tests to work with spatial structure changes
